+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Version 10.0.0 (2023-11-03)
 
+### Features
+
+- new `DoubleRightIconSAComponent` component, that is animated double right icon component
+
 ### BREAKING CHANGES
 - minimal supported version of `@angular` is `17.0.0`
 - minimal supported version of `tslib` is `2.6.2`